+++ bin/ash/main.ml
@@ -15,15 +15,16 @@ let global_opts debug = { debug }
 let virtle_opts global virtle verbose = { global; virtle; verbose }
 
 let spawn opts ssh systemd_ssh_proxy ro_store_socket nix_store_strategy
-    nix_store_image_size_mib config flake override_inputs name user spaces
-    kernel_serial mount_cwd eval ephemeral attach keep kitty waypipe =
+    nix_store_image_size_mib persist_image_size_mib config flake override_inputs
+    name user spaces kernel_serial mount_cwd eval ephemeral attach keep kitty
+    waypipe =
   Log.set_debug opts.global.debug;
   if keep && not attach then Log.fatal "--keep requires --attach";
   if ephemeral && ((not attach) || keep) then
     Log.fatal "--ephemeral requires --attach and cannot be used with --keep";
   Virtle.spawn ?virtle:opts.virtle ?ssh ?systemd_ssh_proxy ?ro_store_socket
-    ?nix_store_strategy ?nix_store_image_size_mib ?name ?user
-    ~config_path:config ?flake ~override_inputs ~spaces ~kernel_serial
+    ?nix_store_strategy ?nix_store_image_size_mib ?persist_image_size_mib ?name
+    ?user ~config_path:config ?flake ~override_inputs ~spaces ~kernel_serial
     ~mount_cwd ~eval ~ephemeral ~attach ~keep ~kitty ~waypipe
     ~verbose:opts.verbose ()
 
@@ -153,6 +154,17 @@ let nix_store_image_size_arg =
           "Override this VM's image-backed Nix store capacity in MiB. Defaults \
            to [global.nix_store].image_size_mib and is saved in \
            ash-state.toml."
+        ~docv:"MIB")
+
+let persist_image_size_arg =
+  Arg.(
+    value
+    & opt (some int) None
+    & info
+        [ "persist-image-size-mib" ]
+        ~doc:
+          "Override this VM's persist image capacity in MiB. Defaults to \
+           [global.persist].image_size_mib and is saved in ash-state.toml."
         ~docv:"MIB")
 
 let config_arg =
@@ -311,9 +323,10 @@ let spawn_cmd =
     Term.(
       const spawn $ virtle_opts_arg $ ssh_arg $ systemd_ssh_proxy_arg
       $ ro_store_socket_arg $ nix_store_strategy_arg $ nix_store_image_size_arg
-      $ config_arg $ flake_arg $ override_input_arg $ name_arg $ user_arg
-      $ spaces_arg $ kernel_serial_arg $ mount_cwd_arg $ eval_flag
-      $ ephemeral_arg $ attach_flag $ keep_flag $ kitty_flag $ waypipe_flag)
+      $ persist_image_size_arg $ config_arg $ flake_arg $ override_input_arg
+      $ name_arg $ user_arg $ spaces_arg $ kernel_serial_arg $ mount_cwd_arg
+      $ eval_flag $ ephemeral_arg $ attach_flag $ keep_flag $ kitty_flag
+      $ waypipe_flag)
 
 let attach_name_arg =
   Arg.(